requirements.txt
wp-api-client==0.1.0
# Optional.
g4f==7.3.0
